+++ jcl/src/jdk.management/share/classes/module-info.java.extra
@@ -22,6 +22,9 @@
  */
 
 exports com.ibm.lang.management;
-exports openj9.lang.management;
 exports com.ibm.virtualization.management;
+/*[IF CRAC_SUPPORT]*/
+exports jdk.crac.management;
+/*[ENDIF] CRAC_SUPPORT */
+exports openj9.lang.management;
 provides sun.management.spi.PlatformMBeanProvider with com.ibm.lang.management.internal.PlatformMBeanProvider;
